Synchrony - first error fixed: specify # of sims directly
Error was "Error in integer(n) : vector size cannot be NA/NaN". Occurring in line 24 because original code specifies nsims from commandArgs. I assume this is to use the command line to run the code on the cluster & specify # of simulations there. 

Now on to another error - "Error in apply(exp_profit, 2, getBestSpp) : object 'getBestSpp' not found".