Quick fix: link to getBestSpp function instead of package to run locally
Fixes error: "Error in apply(exp_profit, 2, getBestSpp) : object 'getBestSpp' not found" across access_sims.R, interaction_sims.R, and synchrony_sims.R, as the existing code was for loading a package with the getBestSpp. That might still be better, but to move forward for now, doing this quick fix.